Renaming core/wrapper.go package to core instead of websocket
